# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * SCRB Sanket — AI Crime Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  investigator: 'investigator',
  supervisor: 'supervisor',
  admin: 'admin',
} as const;

export interface User {
  id: string;
  name: string;
  role: UserRole;
  district: string;
  badgeNumber?: string;
}

export type ChatMessageInputLanguage = typeof ChatMessageInputLanguage[keyof typeof ChatMessageInputLanguage];


export const ChatMessageInputLanguage = {
  english: 'english',
  kannada: 'kannada',
} as const;

export interface ChatMessageInput {
  message: string;
  language: ChatMessageInputLanguage;
  sessionId?: string;
}

export type ChartDataType = typeof ChartDataType[keyof typeof ChartDataType];


export const ChartDataType = {
  bar: 'bar',
  line: 'line',
  pie: 'pie',
} as const;

export interface ChartDataset {
  label: string;
  data: number[];
}

export interface ChartData {
  type: ChartDataType;
  labels: string[];
  datasets: ChartDataset[];
  title?: string;
}

export interface Hotspot {
  lat: number;
  lng: number;
  intensity: number;
  district: string;
  crimeType: string;
  count: number;
  label?: string;
}

export interface ChatResponse {
  id: string;
  role: string;
  content: string;
  reasoning: string[];
  sources: string[];
  timestamp: string;
  chartData?: ChartData;
  mapData?: Hotspot[];
  firIds?: string[];
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  id: string;
  role: ChatMessageRole;
  content: string;
  language?: string;
  timestamp: string;
  reasoning?: string[];
  sources?: string[];
  chartData?: ChartData;
  mapData?: Hotspot[];
}

export type FirStatus = typeof FirStatus[keyof typeof FirStatus];


export const FirStatus = {
  registered: 'registered',
  under_investigation: 'under_investigation',
  chargesheeted: 'chargesheeted',
  closed: 'closed',
} as const;

export interface Fir {
  id: string;
  firNumber: string;
  district: string;
  policeStation: string;
  crimeType: string;
  subType?: string;
  dateOfIncident: string;
  timeOfIncident?: string;
  status: FirStatus;
  latitude?: number;
  longitude?: number;
  personIds?: string[];
  description?: string;
  isSynthetic?: boolean;
}

export interface FirListResponse {
  data: Fir[];
  total: number;
  limit: number;
  offset: number;
}

export interface StatusStat {
  status: string;
  count: number;
}

export interface FirSummary {
  totalFirs: number;
  openCases: number;
  closedCases: number;
  topCrimeType: string;
  topDistrict: string;
  recentCount: number;
  byStatus?: StatusStat[];
}

export type TrendResponseTrendDirection = typeof TrendResponseTrendDirection[keyof typeof TrendResponseTrendDirection];


export const TrendResponseTrendDirection = {
  rising: 'rising',
  falling: 'falling',
  stable: 'stable',
} as const;

export interface TrendPoint {
  month: string;
  count: number;
  isForecast?: boolean;
}

export interface TrendResponse {
  series: TrendPoint[];
  forecast: TrendPoint[];
  earlyWarning: boolean;
  trendDirection?: TrendResponseTrendDirection;
  percentChange?: number;
}

export interface CrimeTypeStat {
  crimeType: string;
  count: number;
  trend?: string;
}

export interface DistrictStat {
  district: string;
  count: number;
  breakdown?: CrimeTypeStat[];
}

export type EarlyWarningSeverity = typeof EarlyWarningSeverity[keyof typeof EarlyWarningSeverity];


export const EarlyWarningSeverity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface EarlyWarning {
  district: string;
  crimeType: string;
  percentIncrease: number;
  recentCount: number;
  previousCount?: number;
  message: string;
  severity?: EarlyWarningSeverity;
}

export type NetworkNodeType = typeof NetworkNodeType[keyof typeof NetworkNodeType];


export const NetworkNodeType = {
  person: 'person',
  case: 'case',
  location: 'location',
} as const;

export interface NetworkNode {
  id: string;
  label: string;
  type: NetworkNodeType;
  caseCount: number;
  district?: string;
  crimeTypes?: string[];
  group?: number;
}

export type NetworkEdgeRelationship = typeof NetworkEdgeRelationship[keyof typeof NetworkEdgeRelationship];


export const NetworkEdgeRelationship = {
  co_accused: 'co_accused',
  same_mo: 'same_mo',
  shared_location: 'shared_location',
  linked_case: 'linked_case',
} as const;

export interface NetworkEdge {
  source: string;
  target: string;
  relationship: NetworkEdgeRelationship;
  caseId?: string;
}

export interface NetworkGraph {
  nodes: NetworkNode[];
  edges: NetworkEdge[];
}

export interface AuditEntry {
  id: string;
  userId: string;
  userName: string;
  role?: string;
  query: string;
  timestamp: string;
  resultsCount: number;
  ipAddress?: string;
}

export interface District {
  id: string;
  name: string;
  lat: number;
  lng: number;
  stationCount?: number;
}

export type GetFirsParams = {
district?: string;
crimeType?: string;
startDate?: string;
endDate?: string;
limit?: number;
offset?: number;
};

export type GetHotspotsParams = {
crimeType?: string;
startDate?: string;
endDate?: string;
};

export type GetTrendsParams = {
crimeType?: string;
district?: string;
months?: number;
};

export type GetNetworkParams = {
personId?: string;
crimeType?: string;
district?: string;
};

export type GetAuditLogParams = {
limit?: number;
};