# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SCRB Sanket — AI Crime Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chat:
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a chat message and receive AI response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "200":
          description: AI response with supporting data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
...
  /firs:
    get:
      operationId: getFirs
      tags: [firs]
      summary: Get FIR records with optional filters
      parameters:
        - name: district
          in: query
          schema:
            type: string
        - name: crimeType
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: FIR records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FirListResponse"
...
components:
  schemas:
...
    ChatMessageInput:
      type: object
      required: [message, language]
      properties:
        message:
          type: string
        language:
          type: string
          enum: [english, kannada]
        sessionId:
          type: string
    ChatResponse:
      type: object
      required: [id, role, content, reasoning, sources, timestamp]
      properties:
        id:
          type: string
        role:
          type: string
        content:
          type: string
        reasoning:
          type: array
          items:
            type: string
        sources:
          type: array
          items:
            type: string
        timestamp:
          type: string
        chartData:
          $ref: "#/components/schemas/ChartData"
        mapData:
          type: array
          items:
            $ref: "#/components/schemas/Hotspot"
        firIds:
          type: array
          items:
            type: string
...
    ChartData:
      type: object
      required: [type, labels, datasets]
      properties:
        type:
          type: string
          enum: [bar, line, pie]
        labels:
          type: array
          items:
            type: string
        datasets:
          type: array
          items:
            $ref: "#/components/schemas/ChartDataset"
        title:
          type: string
    ChartDataset:
      type: object
      required: [label, data]
      properties:
        label:
          type: string
        data:
          type: array
          items:
            type: number
    Fir:
      type: object
      required: [id, firNumber, district, policeStation, crimeType, dateOfIncident, status]
      properties:
        id:
          type: string
        firNumber:
          type: string
        district:
          type: string
        policeStation:
          type: string
        crimeType:
          type: string
        subType:
          type: string
        dateOfIncident:
          type: string
        timeOfIncident:
          type: string
        status:
          type: string
          enum: [registered, under_investigation, chargesheeted, closed]
        latitude:
          type: number
        longitude:
          type: number
        personIds:
          type: array
          items:
            type: string
        description:
          type: string
        isSynthetic:
          type: boolean
    FirListResponse:
      type: object
      required: [data, total, limit, offset]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Fir"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
...
    Hotspot:
      type: object
      required: [lat, lng, intensity, district, crimeType, count]
      properties:
        lat:
          type: number
        lng:
          type: number
        intensity:
          type: number
        district:
          type: string
        crimeType:
          type: string
        count:
          type: integer
        label:
          type: string
...